Correction path of Entities
