+++ pom.xml
@@ -77,11 +77,7 @@
 	<artifactId>webjars-locator-core</artifactId>
 </dependency>
 
-<dependency>
-	<groupId>org.springframework.boot</groupId>
-	<artifactId>spring-boot-starter-oauth2-client</artifactId>
-</dependency>
-                
+ 
 	</dependencies>
 
 	<build>
