+++ src/main/java/com/petsuite/Services/ServicesApplication.java
@@ -1,6 +1,6 @@
 package com.petsuite.Services;
 
-import com.petsuite.model.DogWalker;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
