+++ src/main/java/com/petsuite/Services/model/Dog.java
@@ -1,4 +1,4 @@
-package com.petsuite.model;
+package com.petsuite.Services.model;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
@@ -7,6 +7,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
@@ -17,33 +18,48 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
-@Table(name = "walk_petition")
+@Table(name = "dog")
 @EntityListeners(AuditingEntityListener.class)
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-public class WalkPetition {
+public class Dog {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer walk_petition_id;
+    private Integer dog_id;
     
     @NotBlank
-    private String walk_petition_time;
+    private String dog_name;
     
     @NotBlank
-    private String walk_petition_address;
+    private String dog_race;
     
-    @NotBlank
-    private String walk_petition_date;
+    @NotNull
+    private float dog_height;
     
-    private String walk_petition_notes;
+    @NotNull
+    private float dog_weight;
     
-    @OneToOne(mappedBy = "client_p")
-    private Client client;
+    @NotNull
+    private Integer dog_age;
+    
+    private String dog_notes;
+    
+    @NotNull
+    private String client_user;
+    
+    @ManyToOne
+    @JoinColumn(name="client_user", nullable=false,updatable = false, insertable = false)
+    private Client client_d;
+    
+    @OneToOne(mappedBy = "dog_wp")
+    private WalkPetition walkPetition;
     
     @OneToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "dog_id", referencedColumnName = "dog_id",updatable = false, insertable = false)
-    private Dog dog_wp;
-       
+    @JoinColumn(name = "walk_invoice_id", referencedColumnName = "walk_invoice_id",updatable = false, insertable = false)
+    private WalkInvoice walkInvoice;
+    
+    
+    
 }
