+++ src/main/java/com/petsuite/Services/model/WalkInvoice.java
@@ -1,6 +1,5 @@
-package com.petsuite.model;
+package com.petsuite.Services.model;
 
-import java.util.Set;
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.EntityListeners;
@@ -9,7 +8,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
@@ -20,44 +18,44 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
-@Table(name = "dog_daycare_invoice")
+@Table(name = "walk_invoice")
 @EntityListeners(AuditingEntityListener.class)
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-public class DogDaycareInvoice {
+public class WalkInvoice {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer dog_daycare_invoice_id;
+    private Integer walk_invoice_id;
     
     @NotBlank
-    private String dog_daycare_invoice_time;
+    private String walk_invoice_time;
     
     @NotBlank
-    private String dog_daycare_invoice_date;
+    private String walk_invoice_address;
+    
+    @NotBlank
+    private String walk_invoice_date;
     
     @NotNull
-    private float dog_daycare_invoice_price;
+    private float walk_invoice_price;
     
-    private boolean dog_daycare_invoice_status;
-    
-    @NotBlank
-    private String dog_daycare_user;
+    private boolean walk_invoice_status;
     
     @NotBlank
     private String client_user;
     
-    @ManyToOne
-    @JoinColumn(name="dog_daycare_user", nullable=false,updatable = false, insertable = false)
-    private DogDaycare dogDaycare;
+    @NotBlank
+    private String dog_walker_user;
     
     @ManyToOne
     @JoinColumn(name="client_user", nullable=false,updatable = false, insertable = false)
-    private Client client_i;
+    private Client client_wi;
     
-    @OneToMany(mappedBy = "dogDaycareInvoice")
-    private Set<DogDaycareService> dogDaycareServices;
+    @ManyToOne
+    @JoinColumn(name="dog_walker_user", nullable=false,updatable = false, insertable = false)
+    private DogWalker dogWalker;
     
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "dog_id", referencedColumnName = "dog_id",updatable = false, insertable = false)
