+++ src/main/java/com/petsuite/Services/repository/ClientRepository.java
@@ -1,6 +1,6 @@
-package com.petsuite.repository;
+package com.petsuite.Services.repository;
 
-import com.petsuite.model.Client;
+import com.petsuite.Services.model.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ClientRepository extends JpaRepository<Client, Integer>{}
