+++ src/main/java/com/petsuite/Services/repository/DogDaycareRepository.java
@@ -1,6 +1,6 @@
-package com.petsuite.repository;
+package com.petsuite.Services.repository;
 
-import com.petsuite.model.DogDaycare;
+import com.petsuite.Services.model.DogDaycare;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DogDaycareRepository extends JpaRepository<DogDaycare, Integer>{}
