+++ src/main/java/com/petsuite/Services/repository/DogDaycareServiceRepository.java
@@ -1,6 +1,6 @@
-package com.petsuite.repository;
+package com.petsuite.Services.repository;
 
-import com.petsuite.model.DogDaycareService;
+import com.petsuite.Services.model.DogDaycareService;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DogDaycareServiceRepository extends JpaRepository<DogDaycareService, Integer>{}
