+++ src/main/java/com/petsuite/Services/repository/DogRepository.java
@@ -1,6 +1,6 @@
-package com.petsuite.repository;
+package com.petsuite.Services.repository;
 
-import com.petsuite.model.Dog;
+import com.petsuite.Services.model.Dog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DogRepository extends JpaRepository<Dog, Integer>{}
