+++ src/main/java/com/petsuite/Services/repository/WalkInvoiceRepository.java
@@ -1,6 +1,6 @@
-package com.petsuite.repository;
+package com.petsuite.Services.repository;
 
-import com.petsuite.model.WalkInvoice;
+import com.petsuite.Services.model.WalkInvoice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WalkInvoiceRepository extends JpaRepository<WalkInvoice, Integer>{}
