+++ src/main/java/com/petsuite/Services/repository/WalkPetitionRepository.java
@@ -1,6 +1,6 @@
-package com.petsuite.repository;
+package com.petsuite.Services.repository;
 
-import com.petsuite.model.WalkPetition;
+import com.petsuite.Services.model.WalkPetition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WalkPetitionRepository extends JpaRepository<WalkPetition, Integer>{}
