+++ src/main/resources/application.properties
@@ -1,12 +1,12 @@
 ## Spring DATASOURCE (DataSourceAutoConfiguration & DataSourceProperties)
-#spring.datasource.url = jdbc:mysql://localhost:3306/pet_suite?autoReconnect=true&useUnicode=true&characterEncoding=UTF-8&allowMultiQueries=true&useSSL=false&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
-#spring.datasource.username = root
-#spring.datasource.password = root
+spring.datasource.url = jdbc:mysql://localhost:3306/pet_suite?autoReconnect=true&useUnicode=true&characterEncoding=UTF-8&allowMultiQueries=true&useSSL=false&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+spring.datasource.username = root
+spring.datasource.password = root
 
 ## Hibernate Properties
 
 # The SQL dialect makes Hibernate generate better SQL for the chosen database
-#spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
 
 # Hibernate ddl auto (create, create-drop, validate, update)
-#spring.jpa.hibernate.ddl-auto = update
+spring.jpa.hibernate.ddl-auto = update
